Task 7 There... - handled long numbers

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
  */
 int print_last_digit(int n)
 {
-	int abs_n = n;
+	long int abs_n = n;
 	int last_n;
 
 	if (n < 0)
